+++ blepro/src/main/java/com/lepu/blepro/ble/LeBp2wBleInterface.kt
@@ -155,7 +155,7 @@ class LeBp2wBleInterface(model: Int): BleInterface(model) {
                 }
 
                 //检查返回是否异常
-                if (bleResponse.pkgType != 0x01.toByte()) {
+                if (bleResponse.pkgType != 0x01) {
                     LepuBleLog.d(tag, "model:$model, fileName = ${fileName}, READ_FILE_START => error")
                     LiveEventBus.get<InterfaceEvent>(InterfaceEvent.LeBP2W.EventLeBp2wReadFileError).post(InterfaceEvent(model, fileName))
                     return
@@ -182,7 +182,7 @@ class LeBp2wBleInterface(model: Int): BleInterface(model) {
                 }
 
                 //检查返回是否异常
-                if (bleResponse.pkgType != 0x01.toByte()) {
+                if (bleResponse.pkgType != 0x01) {
                     LepuBleLog.d(tag, "model:$model, fileName = ${fileName}, READ_FILE_CONTENT => error")
                     LiveEventBus.get<InterfaceEvent>(InterfaceEvent.LeBP2W.EventLeBp2wReadFileError).post(InterfaceEvent(model, fileName))
                     return
@@ -206,7 +206,7 @@ class LeBp2wBleInterface(model: Int): BleInterface(model) {
                     return
                 }
                 //检查返回是否异常
-                if (bleResponse.pkgType != 0x01.toByte()) {
+                if (bleResponse.pkgType != 0x01) {
                     LepuBleLog.d(tag, "model:$model, fileName = ${fileName}, READ_FILE_END => error")
                     LiveEventBus.get<InterfaceEvent>(InterfaceEvent.LeBP2W.EventLeBp2wReadFileError).post(InterfaceEvent(model, fileName))
                     return
@@ -250,7 +250,7 @@ class LeBp2wBleInterface(model: Int): BleInterface(model) {
             //--------------------------写文件--------------------------
             WRITE_FILE_START -> {
                 //检查返回是否异常
-                if (bleResponse.pkgType != 0x01.toByte()) {
+                if (bleResponse.pkgType != 0x01) {
                     LepuBleLog.d(tag, "model:$model, fileName = ${fileName}, WRITE_FILE_START => error")
                     LiveEventBus.get<InterfaceEvent>(InterfaceEvent.LeBP2W.EventLeBp2WriteFileError).post(InterfaceEvent(model, fileName))
                     return
@@ -270,7 +270,7 @@ class LeBp2wBleInterface(model: Int): BleInterface(model) {
             }
             WRITE_FILE_DATA -> {
                 //检查返回是否异常
-                if (bleResponse.pkgType != 0x01.toByte()) {
+                if (bleResponse.pkgType != 0x01) {
                     LepuBleLog.d(tag, "model:$model, fileName = ${fileName}, WRITE_FILE_DATA => error")
                     LiveEventBus.get<InterfaceEvent>(InterfaceEvent.LeBP2W.EventLeBp2WriteFileError).post(InterfaceEvent(model, fileName))
                     return
@@ -298,7 +298,7 @@ class LeBp2wBleInterface(model: Int): BleInterface(model) {
             }
             WRITE_FILE_END -> {
                 //检查返回是否异常
-                if (bleResponse.pkgType != 0x01.toByte()) {
+                if (bleResponse.pkgType != 0x01) {
                     LepuBleLog.d(tag, "model:$model, fileName = ${fileName}, WRITE_FILE_END => error")
                     LiveEventBus.get<InterfaceEvent>(InterfaceEvent.LeBP2W.EventLeBp2WriteFileError).post(InterfaceEvent(model, fileName))
                     return
@@ -322,7 +322,7 @@ class LeBp2wBleInterface(model: Int): BleInterface(model) {
             }
 
             SET_CONFIG -> {
-                if (bleResponse.pkgType != 0x01.toByte()) {
+                if (bleResponse.pkgType != 0x01) {
                     LiveEventBus.get<InterfaceEvent>(InterfaceEvent.LeBP2W.EventLeBp2wSetConfig).post(InterfaceEvent(model, false))
                     LepuBleLog.d(tag, "model:$model,SET_CONFIG => error")
                     return
@@ -344,7 +344,7 @@ class LeBp2wBleInterface(model: Int): BleInterface(model) {
             }
 
             RESET -> {
-                if (bleResponse.pkgType != 0x01.toByte()) {
+                if (bleResponse.pkgType != 0x01) {
                     LiveEventBus.get<InterfaceEvent>(InterfaceEvent.LeBP2W.EventLeBp2wReset).post(InterfaceEvent(model, false))
                     LepuBleLog.d(tag, "model:$model,RESET => error")
                     return
@@ -355,7 +355,7 @@ class LeBp2wBleInterface(model: Int): BleInterface(model) {
             }
 
             FACTORY_RESET -> {
-                if (bleResponse.pkgType != 0x01.toByte()) {
+                if (bleResponse.pkgType != 0x01) {
                     LiveEventBus.get<InterfaceEvent>(InterfaceEvent.LeBP2W.EventLeBp2wFactoryReset).post(InterfaceEvent(model, false))
                     LepuBleLog.d(tag, "model:$model,FACTORY_RESET => error")
                     return
@@ -366,7 +366,7 @@ class LeBp2wBleInterface(model: Int): BleInterface(model) {
             }
 
             FACTORY_RESET_ALL -> {
-                if (bleResponse.pkgType != 0x01.toByte()) {
+                if (bleResponse.pkgType != 0x01) {
                     LiveEventBus.get<InterfaceEvent>(InterfaceEvent.LeBP2W.EventLeBp2wFactoryResetAll).post(InterfaceEvent(model, false))
                     LepuBleLog.d(tag, "model:$model,FACTORY_RESET_ALL => error")
                     return
@@ -377,7 +377,7 @@ class LeBp2wBleInterface(model: Int): BleInterface(model) {
             }
 
             SWITCH_STATE ->{
-                if (bleResponse.pkgType != 0x01.toByte()) {
+                if (bleResponse.pkgType != 0x01) {
                     LiveEventBus.get<InterfaceEvent>(InterfaceEvent.LeBP2W.EventLeBp2wSwitchState).post(InterfaceEvent(model, false))
                     LepuBleLog.d(tag, "model:$model,SWITCH_STATE => error")
                     return
@@ -391,7 +391,7 @@ class LeBp2wBleInterface(model: Int): BleInterface(model) {
             GET_WIFI_ROUTE -> {
                 LepuBleLog.d(tag, "model:$model,GET_WIFI_ROUTE => success")
                 LepuBleLog.d(tag, "model:$model,bytesToHex == " + bytesToHex(bleResponse.content))
-                if (bleResponse.pkgType == 0xFF.toByte()) {
+                if (bleResponse.pkgType == 0xFF) {
                     LiveEventBus.get<InterfaceEvent>(InterfaceEvent.LeBP2W.EventLeBp2WifiScanning).post(InterfaceEvent(model, true))
                 } else {
                     val data = Bp2WifiDevice(bleResponse.content)
@@ -419,7 +419,7 @@ class LeBp2wBleInterface(model: Int): BleInterface(model) {
             }
 
             SET_WIFI_CONFIG -> {
-                if (bleResponse.pkgType != 0x01.toByte()) {
+                if (bleResponse.pkgType != 0x01) {
                     LiveEventBus.get<InterfaceEvent>(InterfaceEvent.LeBP2W.EventLeBp2wSetWifiConfig).post(InterfaceEvent(model, false))
                     LepuBleLog.d(tag, "model:$model,SET_WIFI_CONFIG => error")
                     return
@@ -456,7 +456,7 @@ class LeBp2wBleInterface(model: Int): BleInterface(model) {
                 LepuBleLog.d(tag, "model:$model,GET_USER_LIST_CRC => success")
             }
             DELETE_FILE -> {
-                if (bleResponse.pkgType != 0x01.toByte()) {
+                if (bleResponse.pkgType != 0x01) {
                     LiveEventBus.get<InterfaceEvent>(InterfaceEvent.LeBP2W.EventLeBp2wDeleteFile).post(InterfaceEvent(model, false))
                     LepuBleLog.d(tag, "model:$model,DELETE_FILE => error")
                     return
